+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 #### Fixed
 - Exact match may be deprioritized in the completion list
+- Diagnostic with out-of-spec range generated
 
 ## [1.10.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.9.0...1.10.0) (2023-09-22)